Move dummy data service
