+++ cdisc_rules_engine/services/data_services/__init__.py
@@ -1,10 +1,12 @@
 from .base_data_service import BaseDataService, cached_dataset
 from .data_service_factory import DataServiceFactory
 from .local_data_service import LocalDataService
+from .dummy_data_service import DummyDataService
 
 __all__ = [
     "BaseDataService",
     "cached_dataset",
     "LocalDataService",
     "DataServiceFactory",
+    "DummyDataService"
 ]
