+++ tests/unit/test_dummy_data_service.py
@@ -3,7 +3,7 @@ import pandas as pd
 import pytest
 
 from cdisc_rules_engine.dummy_models.dummy_dataset import DummyDataset
-from cdisc_rules_engine.dummy_services.dummy_data_service import DummyDataService
+from cdisc_rules_engine.services.data_services.dummy_data_service import DummyDataService
 
 
 def test_get_dataset():
